atualização no Gestor Financeiro Pessoal

+++ src/main/java/com/curso/services/DestinatarioService.java
@@ -1,11 +1,14 @@
 package com.curso.services;
 
 
+
 import com.curso.domains.Destinatario;
 import com.curso.domains.Lancamento;
+import com.curso.domains.dtos.CentroCustoDTO;
 import com.curso.domains.dtos.DestinatarioDTO;
 import com.curso.repositories.DestinatarioRepository;
 import com.curso.repositories.LancamentoRepository;
+import com.curso.services.exceptions.DataIntegrityViolationException;
 import com.curso.services.exceptions.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -35,40 +38,26 @@ public class DestinatarioService {
     }
 
     public Destinatario create(DestinatarioDTO dto){
-        Destinatario obj = new Destinatario();
-
-        obj.setRazaoSocial(dto.getRazaoSocial());
-        obj.setValor(dto.getValor());
-        obj.setDataRecibi(dto.getDataRecibi());
-
-        if (dto.getLancamentoId() != null) {
-            Lancamento lancamento = lancamentoRepo.findById(dto.getLancamentoId())
-                    .orElseThrow(() -> new ObjectNotFoundException("Lançamento não encontrado! Id: " + dto.getLancamentoId()));
-            obj.setLancamento(lancamento);
-        }
-
-        return destinatarioRepo.save(obj);
+        dto.setId(null);
+        Destinatario obj = new Destinatario(dto);
+        return  destinatarioRepo.save(obj);
     }
 
     public Destinatario update(Integer id, DestinatarioDTO objDto){
-        Destinatario oldObj = destinatarioRepo.findById(id)
-                .orElseThrow(() -> new ObjectNotFoundException("Destinatário não encontrado! Id: " + id));
-
-        oldObj.setRazaoSocial(objDto.getRazaoSocial());
-        oldObj.setValor(objDto.getValor());
-        oldObj.setDataRecibi(objDto.getDataRecibi());
-
-        if (objDto.getLancamentoId() != null) {
-            Lancamento lancamento = lancamentoRepo.findById(objDto.getLancamentoId())
-                    .orElseThrow(() -> new ObjectNotFoundException("Lançamento não encontrado! Id: " + objDto.getLancamentoId()));
-            oldObj.setLancamento(lancamento);
-        }
-
+        objDto.setId(id);
+        Destinatario oldObj = findById(id);
+        oldObj = new Destinatario(objDto);
         return destinatarioRepo.save(oldObj);
     }
 
     public void delete(Integer id){
         Destinatario obj = findById(id);
+        if (!lancamentoRepo.findById(id).isEmpty()) {
+            throw new DataIntegrityViolationException(
+                    "Não é possível deletar: existem lançamentos vinculados a este Destinatario"
+            );
+        }
         destinatarioRepo.deleteById(id);
     }
 }
+
